+++ app/src/components/submissionPage/SnapList.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Snap from "../../firebase/snaps/Snap";
+import Snap from "../../types/Snap";
 
 interface SnapListProps {
     snaps: Snap[];
