+++ app/src/firebase/users/__mocks__/GetSnappables.ts
@@ -1,0 +1,12 @@
+import Snappable from "../../../types/Snappable";
+
+export default async function GetSnappables(): Promise<Snappable[]> {
+    return [
+        {
+            id: "1",
+            email: "hello@example.com",
+            fullName: "Bob bobson",
+            username: "bb123",
+        },
+    ];
+}
