+++ app/src/types/Snap.ts
